fix: clear leftover frontend react-doctor findings (#353)
Bring the leftover post-#351 doctor set to a clean scan on latest
react-doctor: pure localStorage updates, stable autosave hashing,
DialogueLine/WriteModeView refactors, and documented FP/intentional
suppressions via doctor.config.json.
