+++ apps/frontend/src/components/ide-shared/ZipImportFilesDialog/ZipImportFilesDialogStatus.tsx
@@ -29,7 +29,7 @@ export function ZipImportFilesDialogProgress({
       </div>
       <div className="h-2 bg-muted rounded-full overflow-hidden">
         <div
-          className="h-full bg-primary transition-all duration-300"
+          className="h-full bg-primary transition-[width] duration-300"
           style={{ width: `${importState.progress}%` }}
         />
       </div>
