+++ apps/frontend/src/components/write-mode/TechnicalPopover.tsx
@@ -184,6 +184,7 @@ export function TechnicalPopover({
               </div>
               <ul className="text-xs text-muted-foreground space-y-1.5">
                 {visualsData.map((visual, index) => (
+                  // react-doctor-disable-next-line react-doctor/no-array-index-as-key -- type_target can collide; index is local list order tiebreaker
                   <li
                     key={`${visual.type}_${visual.target}_${index}`}
                     className="flex items-center gap-1.5"
