+++ apps/frontend/src/contexts/ToastContext.tsx
@@ -186,6 +186,7 @@ function ToastContainer({ toasts, onRemove }: ToastContainerProps) {
       className="fixed bottom-4 right-4 z-[100] flex flex-col gap-2 pointer-events-none"
     >
       {toasts.map((toast) => (
+        // react-doctor-disable-next-line react-doctor/no-transition-all -- FP: animate-in uses @keyframes enter (opacity/transform), not transition:all; rule docs say Tailwind classNames are out of scope
         <div
           key={toast.id}
           role={toast.variant === "destructive" ? "alert" : "status"}
