+++ apps/frontend/src/hooks/useLocalStorage.ts
@@ -146,6 +146,10 @@ export function useLocalStorage<T>(
   const [state, setState] = useState<T>(() =>
     readStorageValue(prefixedKey, defaultValue, deserializer, validate, ssrSafe)
   );
+  const stateRef = useRef(state);
+  useEffect(() => {
+    stateRef.current = state;
+  }, [state]);
 
   // Re-read from localStorage when key changes
   useEffect(() => {
@@ -163,16 +167,16 @@ export function useLocalStorage<T>(
 
   const setItem = useCallback(
     (value: SetStateAction<T>) => {
-      setState((previousValue) => {
-        const nextValue =
-          typeof value === "function"
-            ? (value as (prev: T) => T)(previousValue)
-            : value;
+      const nextValue =
+        typeof value === "function"
+          ? (value as (prev: T) => T)(stateRef.current)
+          : value;
 
-        if (ssrSafe && !isStorageAvailable()) {
-          return nextValue;
-        }
+      // Keep ref in sync immediately so chained functional updaters see
+      // the latest value before the next render flushes state.
+      stateRef.current = nextValue;
 
+      if (!(ssrSafe && !isStorageAvailable())) {
         try {
           if (nextValue === undefined) {
             window.localStorage.removeItem(prefixedKey);
@@ -185,9 +189,9 @@ export function useLocalStorage<T>(
         } catch (error) {
           logStorageWarning("save", prefixedKey, error);
         }
+      }
 
-        return nextValue;
-      });
+      setState(nextValue);
     },
     [prefixedKey, ssrSafe]
   );
